feat: achieve full coherence parity for Qwen 3.5 & Gemma 4 on CPU/CUDA
- Fix GatedDeltaNet SSM recurrent state updates, per-head L2 normalization, and causal Conv1d shift for Qwen 3.5
- Implement complete Gemma 4 architecture on CPU/CUDA with sliding window attention, shared KV cache, and per-layer embeddings (PLE)
- Upgrade CUDA backend with corrected GGUF tensor dimensions, F32 RMSNorm strides, persistent dequantization scratch buffers, and memory reclamation (runtime.GC / debug.FreeOSMemory)
- Add hybrid layer offloading with contiguous CPU layer batching and pure CPU/GPU routing
- Support universal row-level dequantization for token embeddings across all GGUF quant types (Q8_0, Q4_K, Q6_K, Q2_K, Q3_K, Q5_K, Q4_0, F16, BF16)
- Add 25-question coherence verification suite and unit tests

+++ internal/simd/avx512.go
@@ -468,7 +468,7 @@ func rmsnormScalar(input, weight, output []float32, rows, cols int, eps float32)
 			v := input[offset+c]
 			sum += v * v
 		}
-		sum = float32(1.0) / (float32(math.Sqrt(float64(sum)/float64(cols)) + float64(eps)))
+		sum = float32(1.0 / math.Sqrt(float64(sum)/float64(cols) + float64(eps)))
 
 		// Normalize and scale
 		for c := 0; c < cols; c++ {
